Moved logging setup from SOAPClient to NammuController.
